refactor: share logical model + plan expansion between the two compilers
amber and the workflow-compiling-service each carried a near-duplicate WorkflowCompiler and
logical-plan model (LogicalPlan, LogicalPlanPojo, LogicalLink) — the duplication was even
flagged in-code ("we should consider merge this compile with WorkflowCompilingService's").
Move the canonical pieces into the shared workflow-operator module:

- org.apache.texera.amber.compiler.model.{LogicalLink, LogicalPlan, LogicalPlanPojo} now live
  in workflow-operator as the single copy (the LogicalLink string serializer from #5042 is
  preserved). The duplicates in amber and workflow-compiling-service are deleted.
- PhysicalPlanExpander.expand holds the logical-to-physical expansion the two compilers shared.
- Each compiler is now a thin wrapper over it: amber's adds result-storage planning and builds
  a runtime Workflow (carrying the logical plan) for execution; the compiling-service's adds
  output-schema collection and error reporting for the editor.

amber's references migrate to the shared model package. No behavior change: the logical-link
round-trip, compiler-produced physical-plan serde, and storage-port collection specs all pass.
