+++ amber/src/test/scala/org/apache/texera/amber/engine/e2e/TestUtils.scala
@@ -59,9 +59,10 @@ import org.apache.texera.dao.jooq.generated.tables.pojos.{
   WorkflowVersion,
   Workflow => WorkflowPojo
 }
-import org.apache.texera.web.model.websocket.request.LogicalPlanPojo
+import org.apache.texera.amber.compiler.model.LogicalPlanPojo
 import org.apache.texera.web.resource.dashboard.user.workflow.WorkflowExecutionsResource.getResultUriByLogicalPortId
-import org.apache.texera.workflow.{LogicalLink, WorkflowCompiler}
+import org.apache.texera.amber.compiler.model.LogicalLink
+import org.apache.texera.workflow.WorkflowCompiler
 
 object TestUtils {
