+++ workflow-compiling-service/src/main/scala/org/apache/texera/amber/compiler/WorkflowCompiler.scala
@@ -28,19 +28,13 @@ import org.apache.texera.amber.compiler.WorkflowCompiler.{
 import org.apache.texera.amber.compiler.model.{LogicalPlan, LogicalPlanPojo}
 import org.apache.texera.amber.core.tuple.Schema
 import org.apache.texera.amber.core.virtualidentity.OperatorIdentity
-import org.apache.texera.amber.core.workflow.{
-  PhysicalLink,
-  PhysicalPlan,
-  PortIdentity,
-  WorkflowContext
-}
+import org.apache.texera.amber.core.workflow.{PhysicalPlan, PortIdentity, WorkflowContext}
 import org.apache.texera.amber.core.workflowruntimestate.FatalErrorType.COMPILATION_ERROR
 import org.apache.texera.amber.core.workflowruntimestate.WorkflowFatalError
 
 import java.time.Instant
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
-import scala.jdk.CollectionConverters.IteratorHasAsScala
 
 object WorkflowCompiler {
   // util function for extracting the error causes
@@ -125,74 +119,6 @@ class WorkflowCompiler(
     context: WorkflowContext
 ) extends LazyLogging {
 
-  // function to expand logical plan to physical plan
-  private def expandLogicalPlan(
-      logicalPlan: LogicalPlan,
-      errorList: Option[ArrayBuffer[(OperatorIdentity, Throwable)]]
-  ): PhysicalPlan = {
-    var physicalPlan = PhysicalPlan(operators = Set.empty, links = Set.empty)
-
-    logicalPlan.getTopologicalOpIds.asScala.foreach { logicalOpId =>
-      val logicalOp = logicalPlan.getOperator(logicalOpId)
-      val allUpstreamLinks = logicalPlan.getUpstreamLinks(logicalOp.operatorIdentifier)
-
-      try {
-        val subPlan = logicalOp.getPhysicalPlan(context.workflowId, context.executionId)
-
-        subPlan
-          .topologicalIterator()
-          .map(subPlan.getOperator)
-          .foreach { physicalOp =>
-            val externalLinks = allUpstreamLinks
-              .filter(link => physicalOp.inputPorts.contains(link.toPortId))
-              .flatMap { link =>
-                physicalPlan
-                  .getPhysicalOpsOfLogicalOp(link.fromOpId)
-                  .find(_.outputPorts.contains(link.fromPortId))
-                  .map(fromOp =>
-                    PhysicalLink(fromOp.id, link.fromPortId, physicalOp.id, link.toPortId)
-                  )
-              }
-
-            val internalLinks = subPlan.getUpstreamPhysicalLinks(physicalOp.id)
-
-            // Add the operator to the physical plan
-            physicalPlan = physicalPlan.addOperator(physicalOp.propagateSchema())
-
-            // Add all the links to the physical plan
-            physicalPlan = (externalLinks ++ internalLinks).foldLeft(physicalPlan) { (plan, link) =>
-              plan.addLink(link)
-            }
-
-            // **Check for Python-based operator errors during code generation**
-            if (physicalOp.isPythonBased) {
-              val code = physicalOp.getCode
-              val exceptionPattern = """#EXCEPTION DURING CODE GENERATION:\s*(.*)""".r
-
-              exceptionPattern.findFirstMatchIn(code).foreach { matchResult =>
-                val errorMessage = matchResult.group(1).trim
-                val error =
-                  new RuntimeException(s"Operator is not configured properly: $errorMessage")
-
-                errorList match {
-                  case Some(list) => list.append((logicalOpId, error)) // Store error and continue
-                  case None       => throw error // Throw immediately if no error list is provided
-                }
-              }
-            }
-          }
-      } catch {
-        case e: Throwable =>
-          errorList match {
-            case Some(list) => list.append((logicalOpId, e)) // Store error
-            case None       => throw e // Throw if no list is provided
-          }
-      }
-    }
-
-    physicalPlan
-  }
-
   /**
     * Compile a workflow to physical plan, along with the schema propagation result and error(if any)
     *
@@ -211,7 +137,7 @@ class WorkflowCompiler(
     logicalPlan.resolveScanSourceOpFileName(Some(errorList))
 
     // 3. expand the logical plan to the physical plan
-    val physicalPlan = expandLogicalPlan(logicalPlan, Some(errorList))
+    val physicalPlan = PhysicalPlanExpander.expand(context, logicalPlan, Some(errorList))
 
     // 4. collect the output schema for each logical op
     // even if error is encountered when logical => physical, we still want to get the input schemas for rest no-error operators
